WeatherAPI/src/app/Component/Authentication/logout/logout.component.ts
import { Component } from '@angular/core';
import { AngularFireAuth } from '@angular/fire/compat/auth';
import { Router } from '@angular/router';

@Component({
  selector: 'app-logout',
  templateUrl: './logout.component.html',
  styleUrls: ['./logout.component.css']
})
export class LogoutComponent {
  constructor(private afAuth: AngularFireAuth, private router: Router) {}

  logout() {
    this.afAuth.signOut()
      .then(() => {
        this.router.navigate(['/login']); 
      })
      .catch((error: any) => {
        console.error('Logout failed:', error);
      });
  }
}